feat: setup Kotest AnnotationSpec
This is a little more interesting than other Kotest spec setups. Because
of how the test.name and namespace.name were always the String literal
entirely and not just the content that made it impossible to set up the
AnnotationSpec since it's a function name and thereby isn't wrapped in
quotes like all the other cases. This change cleans up the quote
wrapping which I think also improves how readable the tests are in the
Neotest summary panel.